# backend/src/main/resources/openapi.yaml
#
# GENERAL INFO
#

openapi: 3.0.0
info:
  title: Hair Salon API
  version: 1.0.0
  contact:
    name: "Irdin Ibisevic"
    email: "irdin.ibisevic@gmail.com"

    #
    # PATHS
    #

paths:
  #------------------------------------------------------------------
  # Hairsalon
  #------------------------------------------------------------------
  /hairsalons:
    #------------------------------------------------------------------
    #
    #------------------------------------------------------------------
    post:
      summary: Create a new hair salon
      operationId: saveHairsalon
      requestBody:
        description: Hair Salon data to be added
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HairsalonInDTO'
      responses:
        '201':
          description: Hair salon created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HairsalonOutDTO'
    #------------------------------------------------------------------
    #
    #------------------------------------------------------------------
  /hairsalons/autocomplete:
    get:
      summary: Receive possible Hairsalon names for autocomplete on landing page
      operationId: autocompleteHairsalonName
      parameters:
        - name: salonNameToComplete
          in: query
          description: name of hairsalon to be resolved
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  type: string
                description: List of hairsalon names
        '400':
          description: Bad Request
        '500':
          description: Internal Server Error

  #------------------------------------------------------------------
  # Single Hairsalon
  #------------------------------------------------------------------
  /hairsalons/{salonId}:
    get:
      summary: Get a single hair salon by ID
      operationId: getHairsalonById
      parameters:
        - name: salonId
          in: path
          description: ID of the hair salon to retrieve
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HairsalonOutDTO'
        '404':
          description: Hair salon not found
        '500':
          description: Internal Server Error

#------------------------------------------------------------------
#components
#------------------------------------------------------------------
components:
  schemas:
    #------------------------------------------------------------------
    # HairsalonInDTO
    #------------------------------------------------------------------
    HairsalonInDTO:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        phoneNumber:
          type: string
      required:
        - name
        - description
        - phoneNumber

    #------------------------------------------------------------------
    # HairsalonOutDTO
    #------------------------------------------------------------------
    HairsalonOutDTO:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        description:
          type: string
        address:
          type: string
        phoneNumber:
          type: string
        businessHours:
          type: array
          items:
            $ref: '#/components/schemas/DailyOpeningHours'
        images:
          type: array
          items:
            $ref: '#/components/schemas/Image'

    #--------------------
    # Image
    #--------------------
    Image:
      type: object
      properties:
        id:
          type: integer
        link:
          type: string
        hairsalon_id:
          type: integer


    #--------------------
    # OpeningHours
    #--------------------
    DailyOpeningHours:
      type: object
      properties:
        day:
          type: string
          enum: [ "Monday", "Tuesday", "Wednesday", "Thursday", "Friday", "Saturday", "Sunday" ]
        open-morning:
          type: string
          format: time
        closing-morning:
          type: string
          format: time
        open-afternoon:
          type: string
          format: time
        closing-afternoon:
          type: string
          format: time


    WeeklyOpeningHours:
      type: array
      items:
        $ref: '#/components/schemas/DailyOpeningHours'

    